+++ src/main/java/chess/controller/ChessController.java
@@ -3,13 +3,13 @@ package chess.controller;
 import chess.domain.Score;
 import chess.domain.result.EndResult;
 import chess.domain.result.MoveResult;
-import chess.dto.ChessPieceDto;
-import chess.dto.CurrentTurnDto;
-import chess.dto.ErrorResponseDto;
-import chess.dto.MoveRequestDto;
-import chess.dto.RoomCreationRequestDto;
-import chess.dto.RoomDeletionRequestDto;
-import chess.dto.RoomResponseDto;
+import chess.dto.request.MoveRequestDto;
+import chess.dto.request.RoomCreationRequestDto;
+import chess.dto.request.RoomDeletionRequestDto;
+import chess.dto.response.ChessPieceDto;
+import chess.dto.response.CurrentTurnDto;
+import chess.dto.response.ErrorResponseDto;
+import chess.dto.response.RoomResponseDto;
 import chess.service.ChessService;
 import chess.service.RoomService;
 import java.net.URI;
