+++ src/main/java/chess/dto/request/MoveRequestDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.request;
 
 import chess.domain.position.Position;
