+++ src/main/java/chess/service/RoomService.java
@@ -3,11 +3,11 @@ package chess.service;
 import chess.dao.RoomDao;
 import chess.domain.GameStatus;
 import chess.domain.chesspiece.Color;
-import chess.dto.CurrentTurnDto;
-import chess.dto.RoomCreationRequestDto;
-import chess.dto.RoomDeletionRequestDto;
-import chess.dto.RoomResponseDto;
-import chess.dto.RoomStatusDto;
+import chess.dto.request.RoomCreationRequestDto;
+import chess.dto.request.RoomDeletionRequestDto;
+import chess.dto.response.CurrentTurnDto;
+import chess.dto.response.RoomResponseDto;
+import chess.dto.response.RoomStatusDto;
 import java.util.List;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
