+++ src/test/java/chess/dao/RoomDaoTest.java
@@ -4,11 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.GameStatus;
 import chess.domain.chesspiece.Color;
-import chess.dto.CurrentTurnDto;
-import chess.dto.RoomResponseDto;
-import chess.dto.RoomStatusDto;
+import chess.dto.response.CurrentTurnDto;
+import chess.dto.response.RoomResponseDto;
+import chess.dto.response.RoomStatusDto;
 import java.util.List;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
