+++ src/test/java/chess/service/RoomServiceTest.java
@@ -5,8 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import chess.dao.RoomDao;
 import chess.domain.GameStatus;
 import chess.domain.chesspiece.Color;
-import chess.dto.RoomCreationRequestDto;
-import chess.dto.RoomDeletionRequestDto;
+import chess.dto.request.RoomDeletionRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
